BISERVER-8503 Separate Security User and Role Mngt webservice from User Role List webservice.
